Refactored remaining lookup functions in model_history, moved vis.py functions to model_history; next step is to turn them into methods and nicely refactor.
