+++ torchlens/user_funcs.py
@@ -6,7 +6,7 @@ from torch import nn
 from torchlens.helper_funcs import warn_parallel
 from torchlens.model_funcs import get_all_submodules, run_model_and_save_specified_activations
 from torchlens.validate import validate_saved_activations_for_model_input
-from torchlens.vis import render_graph
+from model_history import render_graph
 
 
 def get_model_activations(model: nn.Module,
